Move getFilePaths
