--- src/commands/RetrieveSignedPackageCommand.ts
@@ -1,11 +1,0 @@
-import { rokuDeploy } from '../index';
-
-export class RetrieveSignedPackageCommand {
-    async run(args) {
-        await rokuDeploy.retrieveSignedPackage(args.pathToPkg, {
-            host: args.host,
-            password: args.password,
-            outFile: args.outFile
-        });
-    }
-}
